+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Parable PHP ORM
 
+## 0.4.0
+
+_Changes_
+
+- `Database` has gained multiple getters to make it easier to work with.
+- `AbstractRepository` now uses `hasValueSets()` rather than `countValueSets() > 0`.
+- Obsolete soft quoting setting removed, as it was never implemented here. Quoting is done in the `parable-php/query` package.
+- Implemented `__debugInfo()` on `Database` to prevent the password being leaked while being dumped. 
+
+  **NOTE**: This _cannot_ prevent `var_export` usage. It is strongly suggested not to use `var_export` for logging purposes where an instance of `Database` may be logged. 
+
 ## 0.3.0
 
 _Changes_
@@ -11,6 +22,7 @@ _Changes_
 ## 0.2.1
 
 _Changes_
+
 - Entities will no longer attempt to set values on new instances that are already `null`. This makes it possible to more strictly type setters.
 - Repositories gained a method: `findUniqueBy($callable): AbstractEntity`. It returns null on 0, the entity on 1 and throws on 1+.
 
@@ -27,15 +39,18 @@ Revert previous changes since the directory rename actually required a 0.2.0.
 ## 0.1.2
 
 _Changes_
+
 - Renamed `Traits` directory to `Features` since they haven't been traits for a while.
 - Removed `getCreatedAt()` and `getUpdatedAt()` from the `SupportsSomethingedAt` features, since Parable ORM doesn't require them and you should choose how to expose them.
 
 ## 0.1.1
 
 _Changes_
+
 - Remove some debug stuff.
 
 ## 0.1.0
 
 _Changes_
+
 - First release.